requirements.txt
pandas
numpy
scikit-learn==1.3.2 
pymango
python-dotenv==0.21.0
PyYAML
dill
scipy
imblearn ==0.0
xgboost==2.1.3

-e .